FileAPI/FileReader-interface/filereader_error.html

<!DOCTYPE html>
<html>
  <head>
    <meta charset="utf-8">
    <title>FileAPI Test: filereader_abort</title>
    <link rel="author" title="Intel" href="http://www.intel.com">
    <link rel="help" href="http://www.w3.org/TR/FileAPI/#dfn-domerror">
    <link rel="help" href="http://www.w3.org/TR/FileAPI/#abort">
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
  </head>
  <body>
    <div id="log"></div>

    <script>
      var blob = new Blob(["TEST THE ERROR ATTRIBUTE AND ERROR EVENT"]);
      var reader = new FileReader();
      var load = false;

      setup({explicit_done: true});

      test(function() {
        assert_equals(reader.error, null, "The error is null when no error occurred");
      }, "Check if the error attribute is null when no error occurred");

      on_event(reader, "load", function(evt) {
        load = true;
      });

      on_event(reader, "loadend", function(evt) {
        test(function() {
          assert_equals(reader.result, null, "The result is null");
        }, "Check if the result is null when error occurred");

        test(function() {
          assert_false(load, "The load event not occurred");
        }, "Check if the load event not occurred when reads failed");

        done();
      });

      reader.readAsText(blob);
      reader.abort();
    </script>
  </body>
</html>
